+++ js/packages/cli/src/helpers/cache.ts
@@ -16,3 +16,27 @@ export function loadCache(cacheName: string, env: string) {
 export function saveCache(cacheName: string, env: string, cacheContent) {
   fs.writeFileSync(cachePath(env, cacheName), JSON.stringify(cacheContent));
 }
+
+export function updateCandyMachineList(candyAddress: string, env: string) {
+  const fp = `${CACHE_PATH}/${env}_candyMachineList.json`
+  if (!fs.existsSync(fp)){
+      let cList = {
+          candyList: [
+              candyAddress
+          ]
+      }
+      fs.writeFileSync(fp, JSON.stringify(cList));
+  
+  } else {
+      let dataJSON = fs.readFileSync(fp, "utf-8");
+      let data = JSON.parse(dataJSON);
+      data.candyList.unshift(candyAddress);
+      fs.writeFileSync(fp, JSON.stringify(data));
+  }
+}
+export function getCandyMachineList(env: string) {
+  const fp = `${CACHE_PATH}/${env}_candyMachineList.json`;
+  let dataJSON = fs.readFileSync(fp, "utf-8");
+  let data = JSON.parse(dataJSON);
+  return data;
+}
